+++ Italia.Spid.AspNetClassic.WebApp/Global.asax.cs
@@ -1,12 +1,9 @@
 ﻿using Italia.Spid.Authentication.IdP;
 using Newtonsoft.Json;
-using System;
 using System.Collections.Generic;
 using System.Configuration;
 using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
-using System.Web;
 using System.Web.Mvc;
 using System.Web.Optimization;
 using System.Web.Routing;
