+++ Italia.Spid.Authentication/IdP/IdpAuthnResponse.cs
@@ -1,5 +1,5 @@
 ﻿/*
-  Copyright (c) 2017 TEAM PER LA TRASFORMAZIONE DIGITALE
+  Copyright (c) 2017 TPCWare - Nicolò Carandini
 
   This file is licensed to you under the BSD 3-Clause License.
   See the LICENSE file in the project root for more information.
