+++ Italia.Spid.Authentication/OpenId/Placeholder note
@@ -1,0 +1,1 @@
+﻿Placeholder note: All the specific OpenID stuff will be placed in this folder.
